fixed a casting error in my finite FiniteDifference.py

+++ FiniteDifference.py
@@ -10,8 +10,8 @@ def gradients(X, function):
     g = np.empty((len(X),))
     delta = (np.finfo(np.float32).eps)**(1/3)
     for idx in range(len(X)):
-        X_up = np.array([i for i in X])
-        X_dn = np.array([i for i in X])
+        X_up = np.array([i for i in X],dtype='float64') # need to set type as float64 or else the += operation may not work correctly
+        X_dn = np.array([i for i in X],dtype='float64') # need to set type as float64 or else the += operation may not work correctly
         X_up[idx] += delta
         X_dn[idx] -= delta
 
@@ -23,8 +23,8 @@ def hessian(X, function, gradients):
     H = np.empty((len(X),len(X)))
     delta = (np.finfo(np.float32).eps)**(1/3)
     for idx in range(len(X)):
-        X_up = np.array([i for i in X])
-        X_dn = np.array([i for i in X])
+        X_up = np.array([i for i in X],dtype='float64') # need to set type as float64 or else the += operation may not work correctly
+        X_dn = np.array([i for i in X],dtype='float64') # need to set type as float64 or else the += operation may not work correctly
         X_up[idx] += delta
         X_dn[idx] -= delta
         g_up = gradients(X_up, function)
